access_bot/requirements.txt
python-telegram-bot[ext]
aiosqlite
python-dotenv
